+++ reoptjl/views.py
@@ -593,13 +593,27 @@ def get_existing_chiller_default_cop(request):
     return: existing_chiller_cop: default COP of existing chiller [fraction]  
     """
     try:
-        existing_chiller_max_thermal_factor_on_peak_load = float(request.GET['existing_chiller_max_thermal_factor_on_peak_load'])  # need float to convert unicode
-        max_load_kw = float(request.GET['max_load_kw'])
-        max_load_kw_thermal = float(request.GET['max_load_kw_thermal'])
+        existing_chiller_max_thermal_factor_on_peak_load = request.GET.get('existing_chiller_max_thermal_factor_on_peak_load')
+        if existing_chiller_max_thermal_factor_on_peak_load is not None:
+            existing_chiller_max_thermal_factor_on_peak_load = float(existing_chiller_max_thermal_factor_on_peak_load)
+        else: 
+            existing_chiller_max_thermal_factor_on_peak_load = 1.25  # default from REopt.jl
 
-        inputs_dict = {"existing_chiller_max_thermal_factor_on_peak_load": existing_chiller_max_thermal_factor_on_peak_load,
-                        "max_load_kw": max_load_kw,
-                        "max_load_kw_thermal": max_load_kw_thermal}
+        max_load_kw = request.GET.get('max_load_kw')
+        if max_load_kw is not None:
+            max_load_kw = float(max_load_kw)
+
+        max_load_ton = request.GET.get('max_load_ton')
+        if max_load_ton is not None:
+            max_load_kw_thermal = float(max_load_ton) * 3.51685  # kWh thermal per ton-hour
+        else: 
+            max_load_kw_thermal = None
+            
+        inputs_dict = {
+            "existing_chiller_max_thermal_factor_on_peak_load": existing_chiller_max_thermal_factor_on_peak_load,
+            "max_load_kw": max_load_kw,
+            "max_load_kw_thermal": max_load_kw_thermal
+        }
 
         julia_host = os.environ.get('JULIA_HOST', "julia")
         http_jl_response = requests.get("http://" + julia_host + ":8081/get_existing_chiller_default_cop/", json=inputs_dict)